// src/app/domains_dashboard/models/analytics.ts
export interface Custos {
	server1: number;
	server2: number;
	cwp1: number;
	cwp2: number;
	locaweb: number;
	service_id: number;
	created_at: Date;
	created_at_hour: string;
	updated_at: Date;
}

export interface CustosAvg {
	server1: number;
	server2: number;
	cwp1: number;
	cwp2: number;
	locaweb: number;
	id: number;
	created_at: Date;
	created_at_hour: string;
	updated_at: Date;
}